RFDemo/api/requirements.txt
robotframework-Selenium2Library
requests~=2.26.0
selenium
pyopenssl
ndg-httpsclient
pyasn1
requests_toolbelt
pycryptodome
rsa
configparser~=5.2.0
pyyaml~=6.0
jsonpath~=0.82
robotframework-reportportal~=5.2.0
Pillow~=9.0.0
bs4~=0.0.1
pytz
lxml
pyexecjs
urllib3==1.26.8
chardet==4.0.0
redis==3.5.3
redis-py-cluster==2.1.3
robotframework==4.0.3
robotframework-requests==0.8.2
openpyxl~=3.0.9
xlrd~=2.0.1
xlwt~=1.3.0